+++ mrv2/docs/HISTORY.md
@@ -19,6 +19,10 @@ v1.0.9
 - Made circle's center also be drawn with float coordinates.
 - Made pen size go to 1 as minimum.
 - Added back .wmv/a decoders and asf demuxer in TLRENDER_FFMPEG_MINIMAL.
+- Added Video Color Primaries, Color TRC and Color Space FFmpeg information
+  to Media Info Panel's video.
+- Fixed FFmpeg reader to take color coefficients.  Now the gray ramp of
+  Sam Richards is the same.
 
 
 v1.0.8